DataClasses: Followers now contains a Long instead of Int
it is possible for some users/artists to have more followers/subscribers than the max integer value,
so it makes more sense to use a long here

+++ app/src/main/kotlin/ts3musicbot/util/DataClasses.kt
@@ -413,13 +413,13 @@ data class Playability(
 )
 
 data class Followers(
-    val amount: Int = -1,
+    val amount: Long = -1,
 ) {
     override fun toString() = amount.toString()
 
-    fun isEmpty() = amount == -1
+    fun isEmpty() = amount == -1L
 
-    fun isNotEmpty() = amount != -1
+    fun isNotEmpty() = amount != -1L
 }
 
 data class Likes(
